requirements.txt
# Core dependencies for Streamlit Cloud deployment
streamlit>=1.46.0
plotly>=5.0.0
networkx>=3.0
pandas>=1.4.0
requests>=2.27.0
pyyaml>=6.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional: Uncomment if using advanced features
# schedule>=1.2.0
# scikit-learn>=1.0.0
# nltk>=3.8
# spacy>=3.4.0

